feat: add loader to Image loading
