+++ src/shared/components/imageCard/ui/imageCard.tsx
@@ -1,15 +1,21 @@
-import React from 'react'
+import React, { useState } from 'react'
 
 import { clsx } from 'clsx'
 import Image, { type ImageProps } from 'next/image'
 
 import s from './imageCard.module.scss'
 
+import { useFetchLoader } from '@/shared/lib'
+
 type Props = ImageProps & {
   cardClassName?: string
 }
 
 export const ImageCard = ({ src, alt, cardClassName, width, height }: Props) => {
+  const [loading, setLoading] = useState(true)
+
+  useFetchLoader(loading)
+
   return (
     <>
       <Image
@@ -18,6 +24,7 @@ export const ImageCard = ({ src, alt, cardClassName, width, height }: Props) => 
         alt={alt}
         width={width}
         height={height}
+        onLoadingComplete={() => setLoading(false)}
       />
     </>
   )
