+++ src/shared/lib/hooks/index.ts
@@ -3,3 +3,4 @@ export { useGoogleLogin } from './useGoogleLogin'
 export { useAppDispatch, useAppSelector } from './appHooks'
 export { useLoader } from './useLoader'
 export { useFetchLoader } from './useFetchLoader'
+export { useErrorHandler } from './useErrorHandler'
