+++ src/shared/lib/index.ts
@@ -5,6 +5,7 @@ export {
   useGoogleLogin,
   useLoader,
   useFetchLoader,
+  useErrorHandler,
 } from './hooks'
 export { consoleErrors } from './handlers'
 export { axiosBaseQuery } from './axios/axiosBaseQuery'
